requirements.txt
Django==4.2.4
djangorestframework==3.14.0
djoser==2.2.0
drf-yasg==1.21.7
environs==9.5.0
Pillow==10.0.0
psycopg2-binary==2.9.7
pytest==7.4.0
pytest-django==4.5.2
python-opengraph==0.0.2
requests==2.31.0

